EST-888, EST-889 & EST-912: allows temp invoices to be deleted, and will remove paperclip (attachments).
If the temp invoice has not been sent (ie the Document is not attached to anything else), then will also delete the Document itself.

Also, moves the functionality for creating email and postal comms from Document mixins to equivalent mixins on Invoice.  (These just delegate to the Document_xxx mixins under the covers).  The main benefit of this is finer-grained control as to the handling of attachments: when the Document is initially created, it is only attached to the Invoice.  But, when the Invoice is sent (by either Invoice_email or Invoice_print), then additional attachments for the invoice's buyer and seller are created (in addition to that of the comm).

If the temp invoice is deleted after it has been sent, the attachment to the buyer, seller and comm remain (even though the temp invoice's paperclip is deleted along with the temp invoice itself).